+++ packages/client-auth/RELEASES.md
@@ -89,3 +89,13 @@ Based on:
 - [typescript v1.0.0-alpha.8] packages/client-auth
 ### Releases
 - [NPM v1.0.0-alpha.8] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.8 - packages/client-auth
+
+## 2024-07-19 00:09:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.344.0 (2.376.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.9] packages/client-auth
+### Releases
+- [NPM v1.0.0-alpha.9] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.9 - packages/client-auth